+++ CHANGELOG.md
@@ -1,9 +1,12 @@
+## March 10th, 2021 (0.5)
+* New args -quality and -resize for convert.
+
 ## March 10th, 2021 (0.4)
 * New program named as ImageMagick: convert, wich can easily convert a lot of files into another format.
 * New submodule with the structure of programs.
 * Process related functions are now inside process.py.
 
-## November 5th, 2020
+## November 5th, 2020
 * Fixed string errors in python3.
 
 ## January 16th, 2019